refactor(apps): numberOfSeries has uint8_t type
This method always returns a small positive integer.

+++ apps/inference/models/chi2_test.h
@@ -30,7 +30,7 @@ class Chi2Test : public SignificanceTest, public InputTable {
                                       int row) = 0;
 
   // Table
-  int numberOfSeries() const override { return 0; }
+  uint8_t numberOfSeries() const override { return 0; }
 
   void compute() override;
